Blog application commit

+++ server/src/Blogs/Service/BlogsService.ts
@@ -4,15 +4,12 @@ import User from "../../db/models/user";
 import Likes from "../../db/models/likes";
 import { Op } from "sequelize";
 
-
 const ifBlogExists = async (id: number) => {
-  console.log("id is", id);
   const blogExists = await Blogs.findByPk(id);
   return blogExists;
 };
 
 const ifUserExists = async (id: number) => {
-  console.log("id is", id);
   const userExists = User.findByPk(id);
   return userExists;
 };
@@ -20,51 +17,51 @@ const ifUserExists = async (id: number) => {
 export const getAllBlogs = async (req: Request, res: Response) => {
   // get all blogs for all user
   try {
+    const { userId, blogId, items, page, sort } = req.query;
 
+    const filter = userId
+      ? blogId
+        ? { userId, id: blogId }
+        : { userId }
+      : blogId
+      ? { id: blogId }
+      : ({} as any);
+    const sortBy = "createdAt";
+    const sortOrder = sort ? sort : ("DESC" as any);
 
-    const { userId, items, page, sort } = req.query;
-    console.log("req.oarams are", req.query);
-    console.log("userId is", userId);
-    const filter = userId ? { userId } : ({} as any);
-    const sortBy= 'createdAt';
-    const sortOrder= sort? sort : 'DESC' as any
-    
-       const offset = page&&items? (+page-1)*(+items) : 0; 
-    
-   
-    // const allBlogData= await Blogs.findAll({include:'users'});
-    const allBlogsData: Record<string,any> = await Blogs.findAll({
-      where:filter,
-      include: [{ model: User, as: "likedUsers", attributes: ["id", "email"],through:{attributes:[]}}],
-      limit: page&&items? +items:5,
-      offset:offset,
-      order:[[sortBy,sortOrder]]
-     
+    const offset = page && items ? (+page - 1) * +items : 0;
+    const allBlogsData: Record<string, any> = await Blogs.findAll({
+      where: filter,
+      include: [
+        {
+          model: User,
+          as: "likedUsers",
+          attributes: ["id", "email"],
+          through: { attributes: [] },
+        },
+      ],
+      limit: page && items ? +items : 5,
+      offset: offset,
+      order: [[sortBy, sortOrder]],
     });
     res.json({ allBlogsData });
     // if(req.user)
     // {
-      
+
     //   const {id:loggedUserId}= req.user;
-    //   console.log("id is ", loggedUserId);
+    //
     //   const updatedBlogs =Object.values(allBlogsData).forEach((blog:Record<string,any>) => {
     //     const idExists: Boolean = blog.likedUsers.some((user:Record<string,any>) => user.id === loggedUserId);
     //     Object.assign(blog, {liked:idExists})
     //     return blog;
     //   });
-    //   console.log("blogs data ",updatedBlogs);
+    //
     //   res.json({updatedBlogs});
-
-
 
     // }
     // else{
     //   res.json({ allBlogsData });
     // }
-
-    // console.log("all blog data", allBlogData);
-
-    
   } catch (error) {
     console.log("error in getting Blog", error);
     res.status(500).json({ error: "Failed to get blog" });
@@ -78,7 +75,7 @@ export const createBlog = async (req: Request, res: Response) => {
   try {
     const createdBlog = await Blogs.create(blogdata);
     console.log("blog created with blog id is ", createdBlog.id);
-    return res.status(201).json(createBlog);
+    return res.status(201).json(createdBlog);
   } catch (error) {
     console.log("error in creating blog", error);
     res.status(500).json({ error: "Failed to create blog" });
@@ -93,11 +90,11 @@ export const updateBlog = async (req: Request, res: Response) => {
   try {
     const blogExists = await Blogs.findByPk(BlogId);
     if (blogExists) {
-      let { title, description, like } = req.body;
+      let { title, description } = req.body;
       if ((title || description) && blogExists.userId !== loggedUserid) {
-        throw new Error("You are not allowed npt update this blog");
+        throw new Error("You are not allowed not update this blog");
       } else {
-        await blogExists.update({ title, description, likes: like });
+        await blogExists.update({ title, description });
         res.json({ message: "Blog updated succesfully" });
       }
     }
@@ -108,51 +105,56 @@ export const updateBlog = async (req: Request, res: Response) => {
 };
 
 export const likeBlog = async (req: Request, res: Response) => {
-  console.log("req.parma", req.params);
   const blogId = req.params.blogId;
-  console.log("blogid", +blogId);
 
   const { liked } = req.body;
-  console.log("liked is ", liked);  
+
   const loggedUser = req.user as Record<string, any>;
   const { id: loggedUserId } = loggedUser;
   try {
-    console.log("blog id is", blogId);
     const blogExists = await ifBlogExists(+blogId);
+    const obj = {
+      userId: +loggedUserId,
+      blogId: +blogId,
+    };
     if (blogExists && liked) {
-      const obj = {
-        userId: +loggedUserId,
-        blogId: +blogId,
-      };
+      if (await Likes.findOne({ where: obj })) {
+        res.status(400).json("Blog already liked");
+        return;
+      }
       await Likes.create(obj);
       const oldLikes = blogExists.likes;
       await blogExists.update({ likes: oldLikes + 1 });
       res.status(200).json("Blog liked Succesfully");
     } else if (blogExists && !liked) {
-      const deletedRow = await Likes.destroy({
-        where: {
-          [Op.and]: [{ blogId: blogExists.id }, { userId: loggedUserId }],
-        },
-      });
-      const oldLikes = blogExists.likes;
-      if (deletedRow) {
-        await blogExists.update({ likes: oldLikes - 1 });
-        res.status(200).json("Blog disliked Succesfully");
+      if (await Likes.findOne({ where: obj })) {
+        const deletedRow = await Likes.destroy({
+          where: {
+            [Op.and]: [{ blogId: blogExists.id }, { userId: loggedUserId }],
+          },
+        });
+        const oldLikes = blogExists.likes;
+        if (deletedRow) {
+          await blogExists.update({ likes: oldLikes - 1 });
+          res.status(200).json("Blog disliked Succesfully");
+        }
+      } else {
+        res.status(400).json("You cannot dislike the blog");
+        return;
       }
     }
-   
   } catch (error) {
     console.log(error);
-    res.status(500).json("Not able to like the Blog");
+    res.status(500).json("Not able to perform the Action");
   }
 };
 
 export const deleteBlog = async (req: Request, res: Response) => {
   const BlogId = req.params.id;
-  const { id: loggedUserId } = req.user?.id;
+  const loggedUserId = req.user?.id;
   try {
     const blogExists = await Blogs.findByPk(BlogId);
-    if (blogExists && blogExists.id === loggedUserId) {
+    if (blogExists && blogExists.userId === loggedUserId) {
       await blogExists.destroy();
       res.json({ message: "Blog has been deleted succesfully" });
     } else {
